cvprw26/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "disaster_maskrcnn"
version = "0.0.1"
description = ""
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "numpy==1.26.4",
    "albumentations==1.4.24",
    "matplotlib",
    "rasterio",
    "tifffile",
    "scikit-image",
    "torch>=2.0",
    "torchvision>=0.15",
    "pycocotools",
    "faster-coco-eval",
    "pyyaml",
    "opencv-python-headless",
    "supervision",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
